requirements.txt
ctransformers==0.2.5
sentence-transformers>=3.3.1
pinecone-client>=2.2.0
langchain==0.0.225
flask>=3.0.0
pypdf>=3.12.0
python-dotenv>=1.0.0
-e .
